+++ app/src/main/java/com/app/arcabyolimpo/data/repository/product/ProductRepositoryImpl.kt
@@ -106,4 +106,5 @@ class ProductRepositoryImpl @Inject constructor(
     override suspend fun searchProducts(query: String): List<Product> {
         return api.searchProducts(query).map { it.toDomain() }
     }
+
 }